Properties typescript 로 마이그레이션
