+++ src/utils/Properties.ts
@@ -1,3 +1,6 @@
+import fs from 'fs';
+import path from 'path';
+
 export type PropertiesJson = {
   ROOT: string;
   LOGGING: boolean;
@@ -5,8 +8,8 @@ export type PropertiesJson = {
 
   // Client
   SERVER: string;
-  PORT: `${number}`;
-  LAST_UPDATE: string;
+  PORT: `${number}` | number;
+  LAST_UPDATE: string | Date;
   targets: string[];
   ignores: string[];
   DELETE_AFTER_UPLOAD: boolean;
@@ -14,3 +17,48 @@ export type PropertiesJson = {
   // Server
   UPLOAD_DIR: string;
 };
+
+class Properties {
+  public path: string = '';
+  private _prop: PropertiesJson;
+
+  constructor(pathToJson?: string) {
+    if (pathToJson === undefined || pathToJson === null) {
+      if (process.env.TTP_APP_PROPERTIES_FILE_PATH === undefined) {
+        throw new Error(`TTP_APP_PROPERTIES_FILE_PATH env value is undefined`);
+      }
+
+      this.path = path.resolve(
+        __dirname,
+        process.env.TTP_APP_PROPERTIES_FILE_PATH
+      );
+    } else {
+      this.path = pathToJson;
+    }
+
+    this.read();
+  }
+
+  get value() {
+    return this._prop;
+  }
+
+  read() {
+    const propFile = fs.readFileSync(this.path, { encoding: 'utf-8' });
+    if (!propFile) throw 'properties.json is empty!';
+
+    const prop = JSON.parse(propFile);
+    if (!prop.targets.length)
+      throw 'There is no target directories for watching';
+
+    prop.LAST_UPDATE = new Date(prop.LAST_UPDATE);
+    this._prop = prop;
+  }
+
+  write(prop: PropertiesJson) {
+    this._prop = prop;
+    fs.writeFileSync(this.path, JSON.stringify(prop, null, 4));
+  }
+}
+
+export default Properties;
